dashboard pendidikan terakhir

+++ app/Config/Routes.php
@@ -37,39 +37,39 @@ $routes->set404Override();
 // route since we don't have to scan directories.
 $routes->get('/', 'Home::index', ['filter' => 'auth']);
 $routes->get('/login', 'Auth::login', ['filter' => 'authTamu']);
-$routes->get('/pendidikan/rekognisi', 'Pendidikan::rekognisi');
-$routes->get('/pendidikan/rekognisiPendidikanDosen', 'Pendidikan::listRekognisiPendidikan');
-$routes->get('/penelitian/rekognisi', 'Penelitian::rekognisi');
-$routes->get('/penelitian/reportPenelitian', 'Penelitian::reportPenelitian');
-$routes->get('/penelitian/reportJurnal', 'Penelitian::reportJurnal');
-$routes->get('/penelitian/rekognisiPenelitianDosen', 'Penelitian::listRekognisiPenelitian');
-$routes->get('/penelitian/listPenelitianDosen', 'Penelitian::listPenelitianDosen');
-$routes->get('/penelitian/listJurnalDosen', 'Penelitian::listJurnalDosen');
+$routes->get('/pendidikan/rekognisi', 'Pendidikan::rekognisi', ['filter' => 'auth']);
+$routes->get('/pendidikan/rekognisiPendidikanDosen', 'Pendidikan::listRekognisiPendidikan', ['filter' => 'auth']);
+$routes->get('/penelitian/rekognisi', 'Penelitian::rekognisi', ['filter' => 'auth']);
+$routes->get('/penelitian/reportPenelitian', 'Penelitian::reportPenelitian', ['filter' => 'auth']);
+$routes->get('/penelitian/reportJurnal', 'Penelitian::reportJurnal', ['filter' => 'auth']);
+$routes->get('/penelitian/rekognisiPenelitianDosen', 'Penelitian::listRekognisiPenelitian', ['filter' => 'auth']);
+$routes->get('/penelitian/listPenelitianDosen', 'Penelitian::listPenelitianDosen', ['filter' => 'auth']);
+$routes->get('/penelitian/listJurnalDosen', 'Penelitian::listJurnalDosen', ['filter' => 'auth']);
 
-$routes->get('/pengabdian/rekognisi', 'Pengabdian::rekognisi');
-$routes->get('/pengabdian/reportPengabdian', 'Pengabdian::reportPengabdian');
-$routes->get('/pengabdian/reportJurnalPengabdian', 'Pengabdian::reportJurnal');
+$routes->get('/pengabdian/rekognisi', 'Pengabdian::rekognisi', ['filter' => 'auth']);
+$routes->get('/pengabdian/reportPengabdian', 'Pengabdian::reportPengabdian', ['filter' => 'auth']);
+$routes->get('/pengabdian/reportJurnalPengabdian', 'Pengabdian::reportJurnal', ['filter' => 'auth']);
 
-$routes->get('/pengabdian/rekognisiPengabdianDosen', 'Pengabdian::rekognisiPengabdianDosen');
-$routes->get('/pengabdian/listPengabdianDosen', 'Pengabdian::listPengabdianDosen');
-$routes->get('/pengabdian/listJurnalPengabdian', 'Pengabdian::listJurnalPengabdian');
+$routes->get('/pengabdian/rekognisiPengabdianDosen', 'Pengabdian::rekognisiPengabdianDosen', ['filter' => 'auth']);
+$routes->get('/pengabdian/listPengabdianDosen', 'Pengabdian::listPengabdianDosen', ['filter' => 'auth']);
+$routes->get('/pengabdian/listJurnalPengabdian', 'Pengabdian::listJurnalPengabdian', ['filter' => 'auth']);
 
-$routes->get('/gs/reportSitasi', 'Sitasi::index');
-$routes->get('/login', 'Auth::login');
+$routes->get('/gs/reportSitasi', 'Sitasi::index', ['filter' => 'auth']);
+
 $routes->get('/register', 'Auth::register');
-$routes->post('/addDosen', 'Auth::addDosen');
+$routes->post('/addDosen', 'Auth::addDosen', ['filter' => 'auth']);
 $routes->post('/ceklogin', 'Auth::ceklogin');
-$routes->get('/logout', 'Auth::logout');
-$routes->get('/dosen/riwayatPendidikanDosen', 'Dosen::pendidikan');
-$routes->get('/dosen/riwayatJafaDosen', 'Dosen::jafa');
-$routes->get('/dosen/riwayatProfesiDosen', 'Dosen::profesi');
-$routes->get('/dosen/riwayatBnspDosen', 'Dosen::bnsp');
-$routes->get('/dosen/listBnspDosen', 'Dosen::listBnsp');
-$routes->get('/dosen/listProfesiDosen', 'Dosen::listProfesi');
-$routes->get('/dosen/listJafaDosen', 'Dosen::listJafa');
-$routes->get('/dosen/listPendidikanDosen', 'Dosen::listPendidikan');
-$routes->post('/dosen/addPendidikanDosen', 'Dosen::addPendidikanDosen');
-$routes->delete('/dosen/delete/(:num)', 'Dosen::delPendidikanDosen/$1');
+$routes->get('/logout', 'Auth::logout', ['filter' => 'auth']);
+$routes->get('/dosen/riwayatPendidikanDosen', 'Dosen::pendidikan', ['filter' => 'auth']);
+$routes->get('/dosen/riwayatJafaDosen', 'Dosen::jafa', ['filter' => 'auth']);
+$routes->get('/dosen/riwayatProfesiDosen', 'Dosen::profesi', ['filter' => 'auth']);
+$routes->get('/dosen/riwayatBnspDosen', 'Dosen::bnsp', ['filter' => 'auth']);
+$routes->get('/dosen/listBnspDosen', 'Dosen::listBnsp', ['filter' => 'auth']);
+$routes->get('/dosen/listProfesiDosen', 'Dosen::listProfesi', ['filter' => 'auth']);
+$routes->get('/dosen/listJafaDosen', 'Dosen::listJafa', ['filter' => 'auth']);
+$routes->get('/dosen/listPendidikanDosen', 'Dosen::listPendidikan', ['filter' => 'auth']);
+$routes->post('/dosen/addPendidikanDosen', 'Dosen::addPendidikanDosen', ['filter' => 'auth']);
+$routes->delete('/dosen/delete/(:num)', 'Dosen::delPendidikanDosen/$1', ['filter' => 'auth']);
 
 
 /*
